[src] refactor: mv pages into /Pages
