+++ Application/src/App.js
@@ -1,8 +1,8 @@
 import './App.css';
-import LoginPage from './LoginPage';
-import RegistrationPage from './RegistrationPage';
-import DashBoard from './DashBoard.js';
-import NotFound404ErrorPage from './NotFound404ErrorPage';
+import LoginPage from './Pages/LoginPage';
+import RegistrationPage from './Pages/RegistrationPage';
+import DashBoard from './Pages/DashBoard.js';
+import NotFound404ErrorPage from './Pages/NotFound404ErrorPage';
 import {
 	BrowserRouter as Router,
 	Route,
